Merge pids with the same command in the cumulative chart
I think the cumulative chart was supposed to be... cumulative :)
so merge the time of the same process called multiple times.

As a positive effect we show that bootchartd polling for processes
is causing a huge overhead and so better kill it shortly.

Unfortunately it does not help with the chart overlapping the
previous one.