Reduce polling frequency for exited procs
It looks like pidof is the major offender on my freerunner when
running bootchart2, consuming around 30% of the cpu time. This is
a bit more intrusive than what i expect. So sleep 1s instead of
200ms in the loop that checks with pidof proccesses we have to
to wait for.
See http://people.freedesktop.org/~rm/moko/bootchart.png for a
chart.