Fix total cpu time calculation for cumulative graph
Fix the total cpu time or at least make it the same of what the
rest of functions expects. It also matches the cpu graph on
top better, here it was half of the total time but the graph
was preyy much blue.
The old method was using proc_tree.process_tree instead of
proc_tree.process_list which is used in the rest of the function.

While at it remove a couple of FIXME already fixed.