pyboothcartgui: fix output file name to match specified format
Raphael Geissert reports in debian bug #631905:
Even when specifying -f svg, pybootchartgui writes the output to
bootchart.png, which is confusing, and incorrect.

Which is true, we have a default value of bootchart.png for output
file name and we didn't care about the alternative formats.

In order to fix that, remove the default name for the output file
and create it after we have parsed the arguments when we can
take into account the desired format.

This leds to a different behaviour in case of an empty -o parameter
passed to pybootchartgui which in the past loaded the interactive
gui; after the change pybootchartgui must be called explicitly
with -i.