pybootchartgui: fix divide by zero errors in parsing code
If we have more that one entry of data in logs with the same
time we crash in proc_ps and proc_diskstat parsers.

The same issue was reported here:
http://code.google.com/p/pybootchartgui/issues/detail?id=13

xrmx: I've reworked his patch to do all the checks in parsing.py
forcing a minimum value of 1 for the interval between times.

While at it fix another crash in proc_ps parsing triggered by
another tarball attached to same issue.