fix bootchart-collector on kernels with CONFIG_TASKSTATS=n
The function get_family_id was supposed to return a negative number to
indicate that the taskstats API was unavailable. However, its return value was
stored to an unsigned variable, so this condition went undetected in the code.

It now returns zero instead of a negative number, and this appears to fix the
problem.

Signed-off-by: Max Eliaser <max.eliaser@intel.com>