collector: properly zero netlink_recvrom buffer
gcc warns that we may want to pass a fixed size to memset
instead of calling sizeof of the src, it is right to suggest so
since src here is a pointer to a char.
Move the buffer initialization in the caller where the stack
size of it is known.