pybootchartgui: Support labeling custom annotations
Adds support for specifying custom annotations either individually
through the command line --annotate option, or in bulk through an input
file.  This can be useful for decorating the bootchart with other
meaningful time-based data not directly associated with the starting and
stopping of a process.

When providing --annotate with a filename, the file should be formatted
as follows:

<label>  <time>
<label2> <time2>

where <time> is in the format HH:MM:SS, and will be used to calculate a
position in the graph based on the bootchart ending time (displayed in
the graph's header).  <label> will appear above the annotation in the
process tree graph (above the seconds header label).  If more than one
label is given for a specific <time>, these will be concatenated and
separated by '/'.

This change also supports adding similarly formatted <time> values
directly on the --annotate option along with process names, though not
at the same time as a filename.

Moves the ProcessTree instantiation above the annotation processing to
take advantage of proc_tree.duration when calculating where <time>
should appear on the bootchart.

Signed-off-by: Kyle Russell <bkylerussell@gmail.com>